graphql_api/requirements.txt
# GraphQL dependencies
strawberry-graphql[fastapi]==0.219.0
fastapi==0.109.0
hypercorn==0.16.0
pydantic==2.5.3

# Monitoring
prometheus-client==0.19.0

# ML Model (shared with other APIs)
sentence-transformers==2.3.1
torch==2.1.2
transformers==4.36.2
